improve error
2001-12-04  Havoc Pennington  <hp@redhat.com>

	* gconf/gconf-internals.c (open_empty_locked_file): improve error

	* gconf/gconftool.c (main): complain about using --config-source
	but not --direct

	* backends/xml-backend.c (resolve_address): don't try to lock
	readonly sources; fixes the case where users install to a prefix
	that they can write to, was causing problems for people using
	build scripts.

	* gconf/gconfd.c (gconf_set_exception): add missing value to case
	statement so we don't crash

	* examples/basic-gconf-app.c: update comment to say that the
	example isn't really right anymore

	* gconf/gconftool.c (process_schema_list): make sure we have an
	element instead of a text node before we try to process it; kills
	a bunch of irritating warnings.

	* Makefile.am (SUBDIRS): remove standard-schemas subdir, should
	have long ago

+++ backends/xml-backend.c
@@ -278,6 +278,9 @@ resolve_address (const gchar* address, GError** err)
   GConfLock* lock = NULL;
   guint dir_mode = 0700;
   guint file_mode = 0600;
+  gchar** address_flags;
+  gchar** iter;
+  gboolean force_readonly;
   
   root_dir = gconf_address_resource(address);
 
@@ -315,26 +318,51 @@ resolve_address (const gchar* address, GError** err)
             }
         }
     }
+
+  force_readonly = FALSE;
   
+  address_flags = gconf_address_flags (address);  
+  if (address_flags)
+    {
+      iter = address_flags;
+      while (*iter)
+        {
+          if (strcmp (*iter, "readonly") == 0)
+            {
+              force_readonly = TRUE;
+              break;
+            }
+
+          ++iter;
+        }
+    }
+
+  g_strfreev (address_flags);
+
   {
     /* See if we're writable */
-    gboolean writable = FALSE;
+    gboolean writable;
     int fd;
     gchar* testfile;
 
-    testfile = g_strconcat(root_dir, "/.testing.writeability", NULL);    
+    writable = FALSE;
     
-    fd = open(testfile, O_CREAT|O_WRONLY, S_IRWXU);
-
-    if (fd >= 0)
+    if (!force_readonly)
       {
-        writable = TRUE;
-        close(fd);
-      }
+        testfile = g_strconcat(root_dir, "/.testing.writeability", NULL);    
         
-    unlink(testfile);
-
-    g_free(testfile);
+        fd = open(testfile, O_CREAT|O_WRONLY, S_IRWXU);
+        
+        if (fd >= 0)
+          {
+            writable = TRUE;
+            close(fd);
+          }
+        
+        unlink(testfile);
+        
+        g_free(testfile);
+      }
     
     if (writable)
       flags |= GCONF_SOURCE_ALL_WRITEABLE;
